sGIM for the binary outcome

This function can integrate summary information from external source

information.borrowing_lgim_binary(beta_initial,gamma_initial,
theta_ext_ini,theta_ext_est,v_ext,y_int,x_int,used_index)

Arguments

beta_initial

the initial values of the main parameter vector

gamma_initial

the initial values of the shift parameter

theta_ext_ini

the initial values of parameters in the reduced model

theta_ext_est

the summary information (parameter estimates) from the external source

v_ext

the variance-covariance matrix from the external source

y_int

the outcome vector from the internal data

x_int

the covariate matrix from the internal data

used_index

index for parameters in the reduced model in relation to the full model

Value

A list of parameter estimates